refactor(api): split HapiErrorMapper and ErrorMappingRegistry
