+++ api/src/shared/application/errors/error-manager.js
@@ -1,6 +1,3 @@
-import Joi from 'joi';
-import _ from 'lodash';
-
 import { AdminMemberError } from '../../../authorization/domain/errors.js';
 import { ChallengeAlreadyAnsweredError } from '../../../certification/evaluation/domain/errors.js';
 import { CertificateGenerationError } from '../../../certification/results/domain/errors.js';
@@ -17,21 +14,15 @@ import {
 } from '../../../organizational-entities/domain/errors.js';
 import { ArchivedCampaignError, DeletedCampaignError } from '../../../prescription/campaign/domain/errors.js';
 import { CampaignParticipationDeletedError } from '../../../prescription/campaign-participation/domain/errors.js';
-import { AggregateImportError, SiecleXmlImportError } from '../../../prescription/learner-management/domain/errors.js';
+import { SiecleXmlImportError } from '../../../prescription/learner-management/domain/errors.js';
 import { OrganizationCantGetPlacesStatisticsError } from '../../../prescription/organization-place/domain/errors.js';
 import {
   AlreadyAcceptedOrCancelledInvitationError,
   UserHasNoOrganizationMembershipError,
   UserNotMemberOfOrganizationError,
 } from '../../../team/domain/errors.js';
-import { config } from '../../config.js';
 import * as SharedDomainErrors from '../../domain/errors.js';
-import { getBaseLocale } from '../../domain/services/locale-service.js';
-import { getI18n } from '../../infrastructure/i18n/i18n.js';
-import { getChallengeLocale } from '../../infrastructure/utils/request-response-utils.js';
 import { HttpErrors } from './http-errors.js';
-
-const NOT_VALID_RELATIONSHIPS = ['externalId', 'participantExternalId'];
 
 const NOT_FOUND_ERRORS = [
   SharedDomainErrors.CertificationCandidateNotFoundError,
@@ -172,7 +163,7 @@ const INTERNAL_SERVER_ERRORS = [LLMDomainErrors.IncorrectMessagesOrderingError];
 
 const PAYLOAD_TOO_LARGE_ERRORS = [LLMDomainErrors.TooLargeMessageInputError];
 
-function _mapToHttpError(error) {
+export function mapToHttpError(error) {
   // Special cases: hardcoded messages or non-standard patterns
   if (error instanceof SharedDomainErrors.UserNotAuthorizedToAccessEntityError) {
     return new HttpErrors.ForbiddenError('Utilisateur non autorisé à accéder à la ressource');
@@ -249,144 +240,4 @@ function _mapToHttpError(error) {
     return new HttpErrors.PayloadTooLargeError(error.message, error.code, error.meta);
 
   return new HttpErrors.BaseHttpError(error.message);
-}
-
-export class ErrorRegistry {
-  #registry = [];
-
-  #mappingSchema = Joi.array().items(
-    Joi.object({
-      name: Joi.string().required(),
-      httpErrorFn: Joi.function().required(),
-    }),
-  );
-
-  register(mapping) {
-    Joi.assert(mapping, this.#mappingSchema);
-
-    mapping.forEach(({ name, httpErrorFn }) => {
-      if (this.#exists(name)) {
-        throw new Error(`Error ${name} already mapped`);
-      }
-      this.#registry[name] = httpErrorFn;
-    });
-  }
-
-  mapToError(ErrorClass) {
-    if (this.#registry[ErrorClass.name]) {
-      return this.#registry[ErrorClass.name](ErrorClass);
-    }
-  }
-
-  #exists(name) {
-    return this.#registry[name] && config.environment !== 'test';
-  }
-}
-
-export class ErrorHapiManager {
-  #registry;
-
-  constructor(registry) {
-    this.#registry = registry;
-  }
-
-  handle(request, h) {
-    const { response } = request;
-
-    if (response instanceof HttpErrors.BaseHttpError) {
-      return HttpErrors.sendJsonApiError(response, h);
-    }
-
-    if (response instanceof SharedDomainErrors.DomainError) {
-      if (response instanceof SharedDomainErrors.EntityValidationError) {
-        const locale = getChallengeLocale(request);
-        const language = getBaseLocale(locale);
-
-        const jsonApiError =
-          response.invalidAttributes?.map((invalidAttribute) =>
-            this.#formatInvalidAttribute(language, response.meta, invalidAttribute),
-          ) ?? new HttpErrors.InvalidEntityError();
-
-        return HttpErrors.sendJsonApiError(jsonApiError, h);
-      }
-
-      const httpError = this.#registry.mapToError(response) ?? _mapToHttpError(response);
-
-      if (response instanceof AggregateImportError) {
-        httpError.meta.forEach((error) => {
-          error.status = httpError.status;
-        });
-        return HttpErrors.sendJsonApiError(httpError.meta, h);
-      }
-
-      return HttpErrors.sendJsonApiError(httpError, h);
-    }
-    return h.continue;
-  }
-
-  #formatInvalidAttribute(locale, meta, { attribute, message }) {
-    if (!attribute) {
-      return this.#formatUndefinedAttribute({ message, locale, meta });
-    }
-    if (attribute.endsWith('Id') && !NOT_VALID_RELATIONSHIPS.includes(attribute)) {
-      return this.#formatRelationship({ attribute, message, locale, meta });
-    }
-    return this.#formatAttribute({ attribute, message, locale, meta });
-  }
-
-  #translateMessage(locale, key) {
-    const i18n = getI18n(locale);
-    if (!key) return key;
-
-    // use regexp to remove i18n key special chars from key
-    const i18nKey = `entity-validation-errors.${key}`.replace(/[:{}%]/g, '');
-    const translation = i18n.__(i18nKey);
-
-    // when the i18n key is returned, so the translation does not exist
-    if (translation === i18nKey) return key;
-
-    return translation;
-  }
-
-  #formatUndefinedAttribute({ message, locale, meta }) {
-    const errorContent = {
-      message: this.#translateMessage(locale, message),
-      code: undefined,
-      meta,
-      source: undefined,
-      title: 'Invalid data attributes',
-    };
-
-    return new HttpErrors.InvalidEntityError(errorContent);
-  }
-
-  #formatRelationship({ attribute, message, locale, meta }) {
-    const relationship = attribute.replace('Id', '');
-
-    const errorContent = {
-      message: this.#translateMessage(locale, message),
-      code: undefined,
-      meta,
-      source: {
-        pointer: `/data/relationships/${_.kebabCase(relationship)}`,
-      },
-      title: `Invalid relationship "${relationship}"`,
-    };
-
-    return new HttpErrors.InvalidEntityError(errorContent);
-  }
-
-  #formatAttribute({ attribute, message, locale, meta }) {
-    const errorContent = {
-      message: this.#translateMessage(locale, message),
-      code: undefined,
-      meta,
-      source: {
-        pointer: `/data/attributes/${_.kebabCase(attribute)}`,
-      },
-      title: `Invalid data attribute "${attribute}"`,
-    };
-
-    return new HttpErrors.InvalidEntityError(errorContent);
-  }
 }
